refactor: Split nosleep into multiple modules
- 🏗️ Created modular package structure with proper separation of concerns
- 📦 `constants.py`: Windows API constants (`ES_CONTINUOUS`, `ES_SYSTEM_REQUIRED`, etc.)
- 🔧 `core.py`: Main `NoSleep` class with sleep prevention logic
- 🖥️ `cli.py`: Command-line interface with argument parsing and `main()` function
- 📦 `__init__.py`: Package exports (`NoSleep` class and version)
- 🚀 `__main__.py`: Module entry point for `python -m nosleep`
- 🔧 Fixed module shadowing by moving entry script outside package
- ✅ Maintained backward compatibility with same CLI interface
- 🧪 All existing features preserved and tested
- 📝 Updated documentation in module docstrings

The tool can now be used in three ways:
1. `python nosleep.py` (traditional entry point)
2. `python -m nosleep` (module execution)
3. `from nosleep import NoSleep` (programmatic use)

🔧 No functional changes - only structural refactoring
