+++ __main__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+"""
+Entry point for nosleep module
+"""
+
+from .cli import main
+
+if __name__ == "__main__":
+    main()
