+++ constants.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+"""
+Constants for Windows SetThreadExecutionState API
+"""
+
+# Windows API constants for SetThreadExecutionState
+ES_CONTINUOUS = 0x80000000
+ES_SYSTEM_REQUIRED = 0x00000001
+ES_DISPLAY_REQUIRED = 0x00000002
+ES_AWAYMODE_REQUIRED = 0x00000040
